Manifest generation
It will eventually be useful to have a manifest file to easily describe the locations of all the files.

Such capability has been added to `render.py`, since it's already iterating over all the SVGs.

MANIFEST.json is an array of key-value pairs. The keys are a derivative of the SVG source path, and values an array of relative paths to the rendered PNGs and PDFs.